+++ jpf-api/src/main/java/com/joiest/jpf/entity/ShopBatchCouponInterfaceInfo.java
@@ -1,5 +1,6 @@
 package com.joiest.jpf.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class ShopBatchCouponInterfaceInfo {
@@ -42,7 +43,7 @@ public class ShopBatchCouponInterfaceInfo {
     /**
      * 对应豆数量
      */
-    private Integer dou;
+    private BigDecimal dou;
 
     /**
      * 是否激活 0=未激活 1=已激活
@@ -130,11 +131,11 @@ public class ShopBatchCouponInterfaceInfo {
         this.money = money;
     }
 
-    public Integer getDou() {
+    public BigDecimal getDou() {
         return dou;
     }
 
-    public void setDou(Integer dou) {
+    public void setDou(BigDecimal dou) {
         this.dou = dou;
     }
