+++ jpf-domain/src/main/java/com/joiest/jpf/common/po/PayShopProduct.java
@@ -46,9 +46,9 @@ public class PayShopProduct implements Serializable {
     private BigDecimal bid;
 
     /**
-     * 豆
+     * 豆数量
      */
-    private Integer dou;
+    private BigDecimal dou;
 
     /**
      * 产品简介
@@ -186,11 +186,11 @@ public class PayShopProduct implements Serializable {
         this.bid = bid;
     }
 
-    public Integer getDou() {
+    public BigDecimal getDou() {
         return dou;
     }
 
-    public void setDou(Integer dou) {
+    public void setDou(BigDecimal dou) {
         this.dou = dou;
     }
